After algorthm runs, selection now covers all of the text moved
Prior, the selection might not cover the moved text, as the selection was being reset back to its original place and not the new location, however, this didn't cause any bugs, as the algorithm doesn't care if the selection covers the whole line or not, it adjusts all text on any line even being touched by the selection.  So this is not a bug fix, but more of an aesthetic fix.