Fixed logic
There was a mistake of calling `setSelectionToCoverAllMovedText() within the for loop.  This should've never been inside the for loop, an accident from previous refactoring.  I moved it outside of the function, as that function shouldn't be handling that sort of action and because of that, I had to pull out the selection range into the `align` function and add an extra parameter to `adjustSelectedLinesOfText`