A bit better logic implemented here
"Left" is the option that requires no text manipulation, so I've placed it at the top of the list, so a user selecting that will reduce the amount of if-statement checks needed to get to it, since nothing more is happening to that string.  I also removed the currentRowString variable that I was assigning and just directly return the strings within the if-statements, to hopefully reduce some more if-statement checks.